translator-service/requirements.txt
kafka-python==2.0.2
groq
httpx==0.27.0